+++ backend/src/main/java/site/bookmore/bookmore/reviews/dto/ChartRequest.java
@@ -1,10 +1,10 @@
-package site.bookmore.bookmore.books.dto;
+package site.bookmore.bookmore.reviews.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import site.bookmore.bookmore.books.entity.Chart;
+import site.bookmore.bookmore.reviews.entity.Chart;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
