+++ backend/src/main/java/site/bookmore/bookmore/reviews/repository/LikesRepository.java
@@ -1,8 +1,8 @@
-package site.bookmore.bookmore.books.repository;
+package site.bookmore.bookmore.reviews.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.bookmore.bookmore.books.entity.Likes;
-import site.bookmore.bookmore.books.entity.Review;
+import site.bookmore.bookmore.reviews.entity.Likes;
+import site.bookmore.bookmore.reviews.entity.Review;
 import site.bookmore.bookmore.users.entity.User;
 
 import java.util.Optional;
