+++ backend/src/main/java/site/bookmore/bookmore/reviews/repository/ReviewTagRepository.java
@@ -1,9 +1,9 @@
-package site.bookmore.bookmore.books.repository;
+package site.bookmore.bookmore.reviews.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.bookmore.bookmore.books.entity.Review;
-import site.bookmore.bookmore.books.entity.ReviewTag;
-import site.bookmore.bookmore.books.entity.Tag;
+import site.bookmore.bookmore.reviews.entity.Review;
+import site.bookmore.bookmore.reviews.entity.ReviewTag;
+import site.bookmore.bookmore.reviews.entity.Tag;
 
 import java.util.Optional;
