+++ backend/src/main/java/site/bookmore/bookmore/users/service/RanksService.java
@@ -6,8 +6,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
-import site.bookmore.bookmore.books.repository.LikesRepository;
-import site.bookmore.bookmore.books.repository.ReviewRepository;
+import site.bookmore.bookmore.reviews.repository.LikesRepository;
+import site.bookmore.bookmore.reviews.repository.ReviewRepository;
 import site.bookmore.bookmore.users.dto.RanksResponse;
 import site.bookmore.bookmore.users.entity.Ranks;
 import site.bookmore.bookmore.users.entity.User;
