+++ backend/src/main/java/site/bookmore/bookmore/users/service/UserService.java
@@ -1,23 +1,15 @@
 package site.bookmore.bookmore.users.service;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import site.bookmore.bookmore.books.dto.ReviewPageResponse;
-import site.bookmore.bookmore.books.entity.Book;
 import site.bookmore.bookmore.common.exception.conflict.DuplicateEmailException;
 import site.bookmore.bookmore.common.exception.conflict.DuplicateNicknameException;
-import site.bookmore.bookmore.common.exception.not_found.BookNotFoundException;
-import site.bookmore.bookmore.common.exception.not_found.EmailNotFoundException;
 import org.springframework.web.multipart.MultipartFile;
-import site.bookmore.bookmore.common.exception.conflict.DuplicateEmailException;
-import site.bookmore.bookmore.common.exception.conflict.DuplicateNicknameException;
 import site.bookmore.bookmore.common.exception.conflict.DuplicateProfileException;
 import site.bookmore.bookmore.common.exception.not_found.UserNotFoundException;
 import site.bookmore.bookmore.common.exception.unauthorized.InvalidPasswordException;
